day44/kafka-dashboard/requirements.txt
kafka-python>=2.0.2
flask==3.0.3
flask-socketio==5.3.6
python-socketio==5.11.2
pandas==2.2.2
numpy==1.26.4
plotly==5.20.0
gunicorn==22.0.0
pytest==8.2.1
pytest-cov==5.0.0
confluent-kafka==2.4.0
avro-python3==1.10.2
redis==5.0.4
requests==2.31.0
psutil==5.9.8
six>=1.17.0
